processor/requirements.txt

# Python dependencies for the Bookshelf Demo processor.
#
# This project is designed to run locally with minimal external dependencies.
# All packages listed below are open-source and can be installed via pip.

# Filesystem monitoring for event-driven image processing
watchdog>=3.0.0

# Data manipulation and Parquet file generation
pandas>=1.3.0
pyarrow>=10.0.0
requests>=2.28.1
Pillow>=9.0.0

# AWS Bedrock for metadata extraction with Claude 3 Haiku
boto3>=1.26.0